basic/requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
cryptography==3.0
distro==1.5.0
Django==2.2
django-cors-headers==3.5.0
django-debug-toolbar==3.1
django-filter==2.3.0
django-nose==1.4.7
django-rest-swagger==2.2.0
djangorestframework==3.11.1
docker==4.3.0
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==8.4.0
nose==1.3.7
openapi-codec==1.3.2
packaging==20.4
paramiko==2.7.1
pluggy==0.13.1
psycopg2==2.8.4
py==1.9.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
pytest-django==3.10.0
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
texttable==1.6.2
toml==0.10.1
uritemplate==3.0.1
urllib3==1.22
websocket-client==0.57.0
zipp==3.1.0
